chore: move debug api types to common files
